+++ src/docs/HISTORY.md
@@ -39,7 +39,7 @@ v1.4.0
   avoid auto refreshing of thumbnails.  You have to go to any thumbnail and
   RMB->Update->Thumbnails.
 - Added thorough list of Wayland compositors with their version commands.
-- Added Desktop Envionment report if using X11 or Wayland.
+- Added Desktop Envionment report when using X11 or Wayland.
 
 
 v1.3.9